feat: create separate links for style fixes
