+++ src/components/index.tsx
@@ -2,14 +2,14 @@
 export {default as App} from './App';
 
 // Typographies
-export {default as H1} from './shared/typography/Heading1';
-export {default as H2} from './shared/typography/Heading2';
-export {default as H3} from './shared/typography/Heading3';
-export {default as H4} from './shared/typography/Heading4';
-export {default as P1} from './shared/typography/Para1';
-export {default as P2} from './shared/typography/Para2';
-export {default as P3} from './shared/typography/Para3';
-export {default as FancyText} from './shared/typography/FancyText';
+export {default as H1} from './typography/Heading1';
+export {default as H2} from './typography/Heading2';
+export {default as H3} from './typography/Heading3';
+export {default as H4} from './typography/Heading4';
+export {default as P1} from './typography/Para1';
+export {default as P2} from './typography/Para2';
+export {default as P3} from './typography/Para3';
+export {default as FancyText} from './typography/FancyText';
 
 // Marginals
 export {default as ThemeToggle} from './marginals/ThemeToggle';
@@ -20,6 +20,10 @@ export {default as Navbar} from './marginals/Navbar';
 export {default as Footer} from './marginals/Footer';
 
 // Layout
-export {default as Container} from './shared/layout/Container';
-export {default as PageContainer} from './shared/layout/PageContainer';
-export {default as Layout} from './shared/layout/Layout';
+export {default as Container} from './layout/Container';
+export {default as PageContainer} from './layout/PageContainer';
+export {default as Layout} from './layout/Layout';
+
+// Atoms
+export {default as InAppLink} from './links/InAppLink';
+export {default as RedirectLink} from './links/RedirectLink';
