+++ src/components/links/InAppLink.tsx
@@ -1,0 +1,12 @@
+// Libraries
+import {Link} from 'gatsby';
+import styled from 'styled-components';
+
+const InAppLink = styled(Link)`
+  text-decoration: none;
+  font-size: 1.2rem;
+  color: var(--color-primary);
+  font-style: italic;
+`;
+
+export default InAppLink;
