+++ src/components/marginals/Footer.tsx
@@ -12,7 +12,7 @@ import {
 } from '@fortawesome/free-brands-svg-icons';
 
 // Components
-import Container from '../shared/layout/Container';
+import Container from '../layout/Container';
 import {H3, P2, P1} from '..';
 
 // Assets
